Setup unchained launch preparer
